🚚 move `getSession` into session.action file
